Refactor modules and update docs
