+++ pipeline.py
@@ -3,20 +3,20 @@
 import os
 from pathlib import Path
 
-import srt2csv
-import srt2audio
-import correct_times
-import wavs2wav
-import ffmpeg_commands
-import vocabular
-from wavs2wav import (
+import subtitle_csv
+import tts_audio
+import sync_utils
+import audio_utils
+import ffmpeg_utils
+import vocabulary
+from audio_utils import (
     convert_mono_to_stereo,
     normalize_stereo_audio,
     extract_acomponiment_or_vocals,
 )
 
 
-def modify_subtitles(subtitle: Path, vocabular_pth: Path):
+def prepare_subtitles(subtitle: Path, vocabular_pth: Path):
     """Apply vocabulary corrections and return helper paths."""
     directory = subtitle.with_suffix("")
     directory.mkdir(exist_ok=True)
@@ -24,33 +24,33 @@ def modify_subtitles(subtitle: Path, vocabular_pth: Path):
     out_path = directory / f"{subtitle_name}_0_mod.srt"
 
     if not out_path.exists():
-        vocabular.modify_subtitles_with_vocabular_text_only(
+        vocabulary.modify_subtitles_with_vocabular_text_only(
             subtitle, vocabular_pth, out_path
         )
 
     return directory, subtitle_name, out_path
 
 
-def convert_to_csv(directory: Path, subtitle_name: str, out_path: Path, speakers: dict, default_speaker: dict):
+def subtitles_to_audio(directory: Path, subtitle_name: str, out_path: Path, speakers: dict, default_speaker: dict):
     """Convert subtitles to CSV and generate English audio."""
     srt_csv_file = directory / f"{subtitle_name}_1.0_srt.csv"
     if not srt_csv_file.exists():
-        srt2csv.srt_to_csv(out_path, srt_csv_file)
+        subtitle_csv.srt_to_csv(out_path, srt_csv_file)
 
     output_csv_with_speakers = directory / f"{subtitle_name}_1.5_output_speakers.csv"
     if not output_csv_with_speakers.exists():
-        srt2csv.add_speaker_columns(srt_csv_file, output_csv_with_speakers)
+        subtitle_csv.add_speaker_columns(srt_csv_file, output_csv_with_speakers)
 
     output_with_preview_speeds_csv = directory / f"{subtitle_name}_3.0_output_speed.csv"
     if not output_with_preview_speeds_csv.exists():
-        srt2csv.add_speed_columns_with_speakers(
+        subtitle_csv.add_speed_columns_with_speakers(
             output_csv_with_speakers, speakers, output_with_preview_speeds_csv
         )
 
-    if not srt2audio.F5TTS.all_segments_in_folder_check(
+    if not tts_audio.F5TTS.all_segments_in_folder_check(
         output_with_preview_speeds_csv, directory
     ):
-        f5tts = srt2audio.F5TTS()
+        f5tts = tts_audio.F5TTS()
         f5tts.generate_from_csv_with_speakers(
             output_with_preview_speeds_csv,
             directory,
@@ -61,13 +61,13 @@ def convert_to_csv(directory: Path, subtitle_name: str, out_path: Path, speakers
 
     corrected_time_output_speed_csv = directory / f"{subtitle_name}_4_corrected_output_speed.csv"
     if not corrected_time_output_speed_csv.exists():
-        correct_times.correct_end_times_in_csv(
+        sync_utils.correct_end_times_in_csv(
             directory, output_with_preview_speeds_csv, corrected_time_output_speed_csv
         )
 
     output_audio_file = directory / f"{subtitle_name}_5.0_output_audiotrack_eng.wav"
     if not output_audio_file.exists():
-        wavs2wav.collect_full_audiotrack(
+        audio_utils.collect_full_audiotrack(
             directory, corrected_time_output_speed_csv, output_audio_file
         )
 
@@ -78,7 +78,7 @@ def convert_to_csv(directory: Path, subtitle_name: str, out_path: Path, speakers
     return srt_csv_file, stereo_eng_file
 
 
-def handle_video(
+def process_video_file(
     video_path: str,
     directory: Path,
     subtitle_name: str,
@@ -92,8 +92,8 @@ def handle_video(
 
     out_ukr_wav = directory / f"{subtitle_name}_5.5_out_ukr.wav"
     if not out_ukr_wav.exists():
-        command = ffmpeg_commands.extract_audio(video_path, out_ukr_wav)
-        ffmpeg_commands.run(command)
+        command = ffmpeg_utils.extract_audio(video_path, out_ukr_wav)
+        ffmpeg_utils.run(command)
 
     acomponiment = directory / f"{subtitle_name}_5.7_accompaniment_ukr.wav"
     if not acomponiment.exists():
@@ -105,9 +105,9 @@ def handle_video(
 
     output_audio = directory / f"{subtitle_name}_6_out_reduced_ukr.wav"
     if not output_audio.exists():
-        volume_intervals = ffmpeg_commands.parse_volume_intervals(srt_csv_file)
+        volume_intervals = ffmpeg_utils.parse_volume_intervals(srt_csv_file)
         normalize_stereo_audio(out_ukr_wav, out_ukr_wav)
-        ffmpeg_commands.adjust_stereo_volume_with_librosa(
+        ffmpeg_utils.adjust_stereo_volume_with_librosa(
             out_ukr_wav, output_audio, volume_intervals, coef, acomponiment
         )
 
@@ -120,13 +120,13 @@ def handle_video(
             return
 
     if not mix_video.exists():
-        command = ffmpeg_commands.create_ffmpeg_mix_video_file_command(
+        command = ffmpeg_utils.create_ffmpeg_mix_video_file_command(
             video_path, output_audio, stereo_eng_file, mix_video
         )
-        ffmpeg_commands.run(command)
+        ffmpeg_utils.run(command)
 
 
-def make_video_from(
+def create_video_with_english_audio(
     video_path: str,
     subtitle: Path,
     speakers: dict,
@@ -134,16 +134,16 @@ def make_video_from(
     vocabular_pth: Path,
     coef: float,
 ):
-    directory, subtitle_name, out_path = modify_subtitles(subtitle, vocabular_pth)
-    srt_csv_file, stereo_eng_file = convert_to_csv(
+    directory, subtitle_name, out_path = prepare_subtitles(subtitle, vocabular_pth)
+    srt_csv_file, stereo_eng_file = subtitles_to_audio(
         directory, subtitle_name, out_path, speakers, default_speaker
     )
-    handle_video(
+    process_video_file(
         video_path, directory, subtitle_name, srt_csv_file, stereo_eng_file, coef
     )
 
 
-def fast_rglob(root_dir: str | Path, extension: str, exclude_ext: str):
+def list_subtitle_files(root_dir: str | Path, extension: str, exclude_ext: str):
     """Recursively search for files with the given extension, excluding modified files."""
     ext = extension.lstrip('.')
     sbt_files = []
