+++ vocabulary.py
@@ -41,38 +41,6 @@ def parse_vocabular_file(vocabular_path):
     return replacements
 
 
-def modify_subtitles_with_vocabular_wholefile_even_partishally(subtitle_path, vocabular_path, output_path):
-    replacements = parse_vocabular_file(vocabular_path)
-
-    with open(subtitle_path, 'r', encoding='utf-8') as infile:
-        text = infile.read()
-
-    for old, new in replacements:
-        text = text.replace(old, new)
-
-    with open(output_path, 'w', encoding='utf-8') as outfile:
-        outfile.write(text)
-
-    return text
-
-def modify_subtitles_with_vocabular_wholefile(subtitle_path, vocabular_path, output_path):
-    replacements = parse_vocabular_file(vocabular_path)
-
-    with open(subtitle_path, 'r', encoding='utf-8') as infile:
-        text = infile.read()
-
-    # Use regular expressions with word boundaries (\b)
-    for old, new in replacements:
-        # Escape 'old' to avoid special regex characters
-        old_escaped = re.escape(old)
-        # Build the pattern: \bOLD\b ensures we only match full words
-        pattern = rf"\b{old_escaped}\b"
-        text = re.sub(pattern, new, text)
-
-    with open(output_path, 'w', encoding='utf-8') as outfile:
-        outfile.write(text)
-
-    return text
 
 def apply_replacements(line, replacements, whole_words=True):
     """
@@ -87,20 +55,6 @@ def apply_replacements(line, replacements, whole_words=True):
             line = line.replace(old, new)
     return line
 
-def modify_subtitles_with_vocabular(subtitle_path, vocabular_path, output_path):
-    """
-    Reads `subtitle_path` line-by-line, applies the replacements
-    from `vocabular_path`, and writes to `output_path`.
-    """
-    # Get the replacements
-    replacements = parse_vocabular_file(vocabular_path)
-
-    with open(subtitle_path, 'r', encoding='utf-8') as infile, \
-         open(output_path, 'w', encoding='utf-8') as outfile:
-        for line in infile:
-            # Apply the replacements on the current line
-            new_line = apply_replacements(line, replacements)
-            outfile.write(new_line)
 
 def modify_subtitles_with_vocabular_text_only(subtitle_path, vocabular_path, output_path):
     replacements = parse_vocabular_file(vocabular_path)
